% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{nodes_repos_users}
\alias{nodes_repos_users}
\alias{nodes_repos}
\alias{nodes_users}
\alias{edges_user_to_repo}
\alias{edges_repo_to_user}
\alias{edges_common_users}
\alias{edges_common_repos}
\title{Functions for extracting a graph from GitHub API data}
\usage{
nodes_repos_users(oarch)

nodes_repos(oarch, ...)

nodes_users(oarch, ...)

edges_user_to_repo(oarch)

edges_repo_to_user(oarch)

edges_common_users(oarch)

edges_common_repos(oarch)
}
\arguments{
\item{oarch}{Data frame containing data from the GitHub API}

\item{...}{Passed to \code{\link[dplyr:select]{dplyr::select()}}}
}
\description{
Functions for extracting a graph from GitHub API data
}
