% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceatlas2.R
\name{layout.forceatlas2}
\alias{layout.forceatlas2}
\title{ForceAtlas2 layout}
\source{
Jacomy M, Venturini T, Heymann S, Bastian M (2014) ForceAtlas2, a Continuous Graph Layout Algorithm for Handy Network Visualization Designed for the Gephi Software. PLoS ONE 9(6): e98679. \url{https://doi.org/10.1371/journal.pone.0098679}
}
\usage{
layout.forceatlas2(
  graph,
  directed = TRUE,
  iterations = 100,
  linlog = FALSE,
  pos = NULL,
  nohubs = FALSE,
  k = 400,
  gravity = 1,
  ks = 0.1,
  ksmax = 10,
  delta = 1,
  center = NULL,
  tolerance = 0.1,
  dim = 2,
  plotstep = 10,
  plotlabels = TRUE
)
}
\arguments{
\item{graph}{An igraph network or a data frame of three columns: source, target, and weights.}

\item{directed}{Logical. TRUE if the network is directed. Ignored if graph is an igraph object.}

\item{iterations}{Number of iterations to be performed.}

\item{linlog}{Logical. If TRUE the algorithm uses logarithmic attraction force \code{F <- log (1+F)}}

\item{pos}{A data frame or matrix (NumberOfNodes x dimension) of the initial locations of points. If NULL the initial positions are random.}

\item{nohubs}{Logical. If TRUE nodes with high indegree have more central position than nodes with outdegree (for directed graphs).}

\item{k}{Is the repel constant: the greater the constant k the stronger the repulsion force between points.}

\item{gravity}{Gravity constant: indicates how strongly the nodes should be attracted to the center of gravity.}

\item{ks}{Speed constant: the greater the value of \code{ks} the more movement the nodes make under the acting forces.}

\item{ksmax}{Limits the speed from above.}

\item{delta}{Modify attraction force, the weights are raised to the power of \code{delta}}

\item{center}{Center of gravity.}

\item{tolerance}{Tolerance to swinging constant.}

\item{dim}{Dimension of the positions.}

\item{plotstep}{is the frequency of plotting intermediate iterations.}

\item{plotlabels}{Logical. If TRUE the labels should be included in the intermediate interations plot}
}
\value{
When graph is an igraph object, the function returns a matrix with the positions. If graph is a data frame, this is a data frame with the names of the nodes and the positions.
}
\description{
Copied from \url{https://github.com/analyxcompany/ForceAtlas2} (licensed GPL>=2)
}
\examples{
library(igraph)
g <- graph.ring(50)
layout <- layout.forceatlas2(g, iterations=3000, plotstep=100)
plot(g, layout=layout)

}
