% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_filter}
\alias{n_filter}
\title{Count number of rows in a filtered data frame}
\usage{
n_filter(.data, ...)
}
\arguments{
\item{.data}{Data frame}

\item{...}{Filters passed to \code{\link[dplyr:filter]{dplyr::filter()}}}
}
\description{
Count number of rows in a filtered data frame
}
